Title: 10개 입력받기, Time: 110ms, Memory: 9MB, Status: Passed - CodeTree

+++ 240912/10개 입력받기/receive-10-inputs.java
@@ -22,7 +22,7 @@ public class Main {
             count++;
         }
 
-        double average = (double) (sum / count);
+        double average = (double) sum / (double) count;
 
         System.out.println(sum + " " + String.format("%.1f", average));
     }
